Refactor synth_device.cpp

+++ src/domain/devices/synth_device.hpp
@@ -256,7 +256,7 @@ private:
     };
 
     std::vector<Voice> m_voices;
-    int m_polyNextVoice = 0;
+    size_t m_polyNextVoice = 0;
 
     mutable std::mt19937 m_rng { std::random_device {}() };
     mutable std::uniform_real_distribution<double> m_phaseDist { 0.0, 1.0 };
